fix(review): PR #603 R5 — EDT 중복테스트 제거, DST 유틸 위젯계층 마이그레이션, 타입 정의
